Move secrets update logic from `crypto.update` to `rotation` package
