+++ app/src/main/AndroidManifest.xml
@@ -117,7 +117,7 @@
             android:stopWithTask="false"
             android:foregroundServiceType="dataSync" />
         <service
-            android:name=".service.security.crypto.update.SecretsUpdateAndroidService"
+            android:name=".service.security.rotation.SecretsUpdateAndroidService"
             android:exported="false"
             android:stopWithTask="false"
             android:foregroundServiceType="dataSync" />
