+++ service/src/main/java/app/notesr/service/security/rotation/SecretsUpdateAndroidServiceStarter.java
@@ -3,11 +3,11 @@
  * SPDX-License-Identifier: MIT
  */
 
-package app.notesr.service.security.crypto.update;
+package app.notesr.service.security.rotation;
 
 import static java.util.Objects.requireNonNull;
 import static app.notesr.core.util.CharUtils.charsToBytes;
-import static app.notesr.service.security.crypto.update.SecretsUpdateAndroidService.EXTRA_CURRENT_STATE;
+import static app.notesr.service.security.rotation.SecretsUpdateAndroidService.EXTRA_CURRENT_STATE;
 
 import android.content.Context;
 import android.content.Intent;
